Refactoring: youtube, sheet types
