+++ hooks/useSheet.ts
@@ -1,4 +1,4 @@
-import { SheetAPIReturntype } from '@/models/sheet/inSheet';
+import { SheetAPIReturntype } from '@/types/inSheet';
 import useSWR from 'swr';
 import { fetcher } from '@/hooks/fetcher';
