+++ pages/api/sheet/index.ts
@@ -1,5 +1,5 @@
 import type { NextApiRequest, NextApiResponse } from 'next';
-import { SheetAPIReturntype } from '@/models/sheet/inSheet';
+import { SheetAPIReturntype } from '@/types/inSheet';
 import { getSheet } from '@/models/sheet';
 import { parseAllData, parseScheduledData } from '@/utils/parseContentSheet';
 import { serverEnvConfig } from '@/configs';
